[CHORE] Clean Up
